requirements.txt
future==0.16.0
olefile==0.44
Pillow==4.3.0
pyglet==1.3.0
